reorganize plonk
