+++ poly-iops/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "poly-iops"
-version = "0.1.4"
+version = "0.1.5"
 authors = ["Findora <engineering@findora.org>"]
 edition = "2021"
 description = "Zei PLONK Poly-IOP protocol"
@@ -10,24 +10,23 @@ name = "poly_iops"
 crate-type = ["rlib"]
 
 [dependencies]
-algebra = { package = "algebra", path = "../algebra" }
-
-byteorder = { version = "^1.2.3", default-features = false, features = ["i128"] }
-crypto = { package = "crypto", path = "../crypto" }
+algebra = { path = "../algebra" }
+crypto = { path = "../crypto" }
+utils = { path = "../utils" }
+bincode = "1.3.1"
+byteorder = { version = "1.2.3", default-features = false, features = ["i128"] }
 custom_error="1.7.1"
 itertools = "0.9.0"
 merlin = "2.0.0"
-num-bigint = { version = "0.3" }
-num-integer = { version = "0.1.43" }
-num-traits = { version = "0.2.12" }
+num-bigint = "0.3"
+num-integer = "0.1.43"
+num-traits = "0.2.12"
 rand = "0.7.3"
 rand_chacha = "0.2.1"
 rand_core = "0.5.1"
+ruc = "1.0"
 serde = { version = "1.0.115", features = ["derive"] }
 serde_derive = "1.0.115"
-utils = { path = "../utils" }
-bincode = "1.3.1"
-ruc = "1.0"
 
 [dev-dependencies]
 serde_json = "1.0"
