+++ poly-iops/src/lib.rs
@@ -6,6 +6,6 @@
 extern crate serde_derive;
 
 pub mod commitments;
-pub mod ioputils;
 pub mod plonk;
 pub mod polynomials;
+pub mod utils;
