--- poly-iops/src/plonk/mod.rs
@@ -1,10 +1,0 @@
-pub mod errors;
-pub mod field_simulation;
-#[allow(non_snake_case)]
-pub(crate) mod plonk_helpers;
-#[allow(non_snake_case)]
-pub mod plonk_setup;
-#[allow(non_snake_case)]
-pub mod protocol;
-pub mod transcript;
-pub mod turbo_plonk_cs;
